% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abun.rich.sepa}
\alias{abun.rich.sepa}
\title{Arthropod abundance and richness data}
\format{
R data.frame
\describe{
  \item{farmName}{the general name of 6 individual farms of three "urbanization categories" (high, medium, low) and the names of the three categories to plot as comparisons to category means in abundance and richness plots}
  \item{meanAbun}{mean abundance of arthropod predators on each farm across all pitfall traps}
  \item{meanRich}{mean richness of arthropod predator families on each farm across all pitfall traps}
  \item{urbanCat}{the urbanization category of each individual farm to be print out in the .html file}

}
}
\usage{
abun.rich.sepa
}
\description{
Dataframe containing arthropod predator abundance and richness data from southeastern Pennsylvania farms.
}
\keyword{datasets}
