% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arth.div.pie.sepa}
\alias{arth.div.pie.sepa}
\title{Arthropod diversity data for pie chart}
\format{
R data.frame
\describe{
  \item{farmName}{the general name of 6 individual farms of three "urbanization categories" (high, medium, low)}
  \item{family}{common name of arthropod predator family}
  \item{meanAbun}{mean abundance of arthropod predators in each family across all pitfall traps}
  \item{prop}{proportion of family abundance over the total number of arthropod predators across all pitfall traps at a farm}

}
}
\usage{
arth.div.pie.sepa
}
\description{
Dataframe containing proportional abundacne of major orders of arthropod predators from total arthropod abundance on southeastern Pennsylvania farms; used to make pie charts.
}
\keyword{datasets}
