% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{bar}
\alias{bar}
\title{Create a bar plot.}
\usage{
bar(df, x, y, ylab, fillLab)
}
\arguments{
\item{df}{DESCRIPTION.}

\item{x}{DESCRIPTION.}

\item{y}{DESCRIPTION.}

\item{ylab}{DESCRIPTION.}

\item{fillLab}{DESCRIPTION.}
}
\value{
DESCRIPTION
}
\description{
Create a bar plot.
}
\examples{
x <- c("a", "b", "c")
y <- c(2, 4, 4)
mydata <- as.data.frame(cbind(x, y))
ylab <- c("y-axis label")
fillLab <- x
bar(df = mydata,
    x = mydata$x,
    y = mydata$y,
    ylab = ylab,
    fillLab = fillLab)
}
