% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curacao.arthropods}
\alias{curacao.arthropods}
\title{curacao.arthropods}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{Site}{name of the focal property, to be used for report title generation (anonymized by letter here).}
  \item{sitetype}{land-use category for the focal property.}
  \item{SR}{species richness of arthropods found at the property.}
  \item{Spiders}{proportion of all arthropods found at the property that were spiders (Araneae)}
  \item{Beetles}{proportion of all arthropods that were beetles (Coleoptera)}
  \item{Flies}{proportion of all arthropods that were true flies (Diptera)}
  \item{True Bugs}{proportion of all arthropods that were true bugs (Hemiptera)}
  \item{Wasps and Ants}{proportion of all arthropods that were wasps and ants (Hymenoptera)}
  \item{Butterflies}{proportion of all arthropods that were butterflies (Lepidoptera)}
  \item{Lacewings}{proportion of all arthropods that were lacewings (Neuroptera)}
  \item{Crickets}{Proportion of all arthropods that were crickets (Orthoptera)}
}
}
\usage{
curacao.arthropods
}
\description{
Some description.
}
\keyword{datasets}
