% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curacao.exo.birds}
\alias{curacao.exo.birds}
\title{curacao.exo.birds}
\format{
\describe{
  \item{Site}{name of the focal property (anonymized by letter here)}
  \item{sitetype}{land-use category for the focal property}
  \item{Exotic parrots}{abundance of exotic parrots at focal property}
  \item{Chestnut-fronted Macaw}{abundance of Chestnut-fronted Macaws at focal property}
  \item{Rock Pigeon}{abundance of Rock Pigeons at focal property}
  \item{House Sparrow}{abundance of House Sparrows at focal property}
  \item{Saffron Finch}{abundance of Saffron Finches at focal property}
  \item{ReportName}{name of report for references (not required)}

}
}
\usage{
curacao.exo.birds
}
\description{
Some description.
}
\keyword{datasets}
