% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curacao.exo.birds.sum}
\alias{curacao.exo.birds.sum}
\title{Curaçao exotic birds summary data}
\format{
R data.frame
\describe{
  \item{species}{common name of exotic bird species observed on properties}
  \item{Site Type}{land-use category for the focal property}
  \item{Percentage of Sites}{The number of sites where each bird species was observed in each category}
}
}
\usage{
curacao.exo.birds.sum
}
\description{
Dataframe summarizing the sites where exotic birds were observed at Curaçao homeowner sites.
}
\keyword{datasets}
