% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curacao.exo.herps}
\alias{curacao.exo.herps}
\title{curacao.exo.herps}
\format{
\describe{
  \item{Site}{name of the focal property (anonymized by letter here)}
  \item{sitetype}{land-use category for the focal property}
  \item{Mourning Gecko}{Abundance of Mourning Gecko at focal property.}
  \item{Tropical House Gecko}{Abundance of Tropical House Gecko at focal property.}
  \item{Johnstone's Whistling Frog}{Abundance of Johnstone's Whistling Frog at focal property}
  \item{Cuban Tree Frog}{Abundance of Cuban Tree Frog at focal property}
  \item{Colombian Four-eyed Frog}{Abundance of Colombian Four-eyed Frog at focal property}
  \item{Asian House Gecko}{Abundance of Asian House Gecko at focal property}
  \item{ReportName}{name of report for references (not required)}

}
}
\usage{
curacao.exo.herps
}
\description{
Some description.
}
\keyword{datasets}
