% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curacao.exo.herps.sum}
\alias{curacao.exo.herps.sum}
\title{Curaçao exotic herps summary data}
\format{
R data.frame
\describe{
  \item{species}{common name of exotic herp species observed on properties}
  \item{Site Type}{land-use category for the focal property}
  \item{Percentage of Sites}{the number of sites where each herp species was observed in each category}

}
}
\usage{
curacao.exo.herps.sum
}
\description{
Dataframe summarizing the sites where exotic reptiles and amphibians were observed at Curaçao homeowner sites.
}
\keyword{datasets}
