% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curacao.nat.birds.list}
\alias{curacao.nat.birds.list}
\title{curacao.nat.bird.list}
\format{
\describe{
  \item{Site}{Name of the focal property (anonymized by letter here)}
  \item{sitetype}{Land-use category for the focal property}
  \item{Scientific name}{The number of sites where each herp species was observed in each category}
  \item{Present}{The number of sites where each herp species was observed in each category}
  \item{Common Name}{Common name of native bird species in Curaçao}

}
}
\usage{
curacao.nat.birds.list
}
\description{
Some description.
}
\keyword{datasets}
