% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curacao.nat.birds}
\alias{curacao.nat.birds}
\title{Curaçao native bird data}
\format{
R data.frame
\describe{
  \item{Site}{name of the focal property (anonymized by letter here)}
  \item{sitetype}{land-use category for the focal property}
  \item{SR}{species richness of native birds found at the focal property}
  \item{ReportName}{name of report for references (not necessary)}

}
}
\usage{
curacao.nat.birds
}
\description{
Dataframe containing native bird data from Curaçao homeowner sampling.
}
\keyword{datasets}
