% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curacao.nat.herps}
\alias{curacao.nat.herps}
\title{Curaçao native herp data}
\format{
R data.frame
\describe{
  \item{Site}{name of the focal propety (anonymized by letter here)}
  \item{sitetype}{land-use category for the focal property}
  \item{Dutch Leaf-toed Gecko}{abundance of Dutch Leaf-toed Gecko at focal property}
  \item{Antillean Dwarf Gecko}{abundance of Antillean Dwarf Gecko at focal property}
  \item{Striped Anole}{abundance of Striped Anole at focal property}
  \item{Laurent's Whiptail}{abundance of Laurent's Whiptail at focal property}
  \item{Turnip-tailed Gecko}{abundance of Turnip-tailed Gecko at focal property}
  \item{ReportName}{name of report for references (not necessary)}

}
}
\usage{
curacao.nat.herps
}
\description{
Dataframe containing native bird data from Curaçao homeowner sampling.
}
\keyword{datasets}
