% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curacao.nat.herps.sum}
\alias{curacao.nat.herps.sum}
\title{curacao.nat.herp.sum}
\format{
\describe{
  \item{sitetype}{Land-use category for the focal property}
  \item{species}{Common name of native herp species on properties}
  \item{mean}{Mean abundance of each native herp species on each property within a land-use category}

}
}
\usage{
curacao.nat.herps.sum
}
\description{
Some description.
}
\keyword{datasets}
