% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interact.rich.sepa}
\alias{interact.rich.sepa}
\title{Plant community & arthropod richness data}
\format{
R data.frame
\describe{
  \item{farmName}{the general name of 15 individual farms of three "urbanization categories" (low, medium, high urbanization		  intensity) and the names of the three categories to plot as comparisons to category means in abundance and richness plots.}
  \item{richness}{richness of arthropod predator families in each pitfall trap on a farm}
  \item{urbanCat}{the urbanization category of each individual farm to be print out in the .html file}
  \item{weedyCover}{the percent ground cover of non-crop plants in a 2-meter radius circle around the pitfall trap}

}
}
\usage{
interact.rich.sepa
}
\description{
Dataframe containing arthropod richness and plant community data from southeastern Pennsylvania farms.
}
\keyword{datasets}
