% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals.each.site}
\alias{mammals.each.site}
\title{Mammals and birds presence/absence}
\format{
R data.frame
\describe{
  \item{Site}{site name}
  \item{site_code}{site code (two-three letter unique identifier when not anonymized)}
  \item{Ownership}{property manager of that site}
  \item{County}{site county}
  \item{Present at your site}{character string of animals present}

}
}
\usage{
mammals.each.site
}
\description{
Dataframe containing names of animals found at each site.
}
\keyword{datasets}
