% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals.photos}
\alias{mammals.photos}
\title{Mammals and birds observation count data}
\format{
R data.frame
\describe{
  \item{Site}{site name}
  \item{site_code}{site code (two-three letter unique identifier when not anonymized)}
  \item{Ownership}{property manager of that site}
  \item{County}{site county}
  \item{species}{animal species}
  \item{species_photos}{number of species' photographed observations}

}
}
\usage{
mammals.photos
}
\description{
Dataframe containing the number of photo observations of animals species at each site.
}
\keyword{datasets}
