% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals.species.richness.county}
\alias{mammals.species.richness.county}
\title{Mammals and birds richness by county data}
\format{
R data.frame
\describe{
  \item{County}{county name}
  \item{county_avg_rich}{average species richness in each county}

}
}
\usage{
mammals.species.richness.county
}
\description{
Dataframe containing mammal and bird species richness by county.
}
\keyword{datasets}
