% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals.species.richness}
\alias{mammals.species.richness}
\title{Mammals and birds richness by site data}
\format{
R data.frame
\describe{
  \item{Site}{site name}
  \item{site_code}{site code (two-three letter unique identifier when not anonymized)}
  \item{Ownership}{property manager of that site}
  \item{County}{site county}
  \item{species_richness}{species richness at each site}
  \item{county_avg_rich}{average species richness in site's county}

}
}
\usage{
mammals.species.richness
}
\description{
Dataframe containing mammal and bird species richness by site.
}
\keyword{datasets}
