% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathogens.county}
\alias{pathogens.county}
\title{Tick-borne disease prevalence by county data}
\format{
R data.frame
\describe{
  \item{County}{county name}
  \item{Disease county}{tick-borne disease combination indicator ('new' does _not_ mean new to that county)}
  \item{nymphs_county}{total tick nymphs tested for disease in that county}
  \item{prop_inf_county}{proportion of county's nymphs found with that combination of tick-borne disease}
  \item{Disease}{disease name to present in figure}
}
}
\usage{
pathogens.county
}
\description{
Dataframe containing tick-borne disease prevalence in tick nymphs.
}
\keyword{datasets}
