% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathogens}
\alias{pathogens}
\title{Tick-borne disease prevalence by site data}
\format{
R data.frame
\describe{
  \item{Site}{site name}
  \item{total_nymphs_site}{tick nymphs tested for disease at that site}
  \item{Disease}{disease name to present in figure}
  \item{Ownership}{property manager of that site}
  \item{prop_inf}{proportion of the site's nymphs found with that combination of tick-borne disease}

}
}
\usage{
pathogens
}
\description{
Dataframe containing arthropod abundance and richness data from traps based on presence/absence of plastic mulch.
}
\keyword{datasets}
