% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathogens.table}
\alias{pathogens.table}
\title{Tick-borne disease presence/absence and tick nymph abundance data}
\format{
R data.frame
\describe{
  \item{Site}{site name}
  \item{total_nymphs_site}{tick nymphs tested for disease at that site}
  \item{Ownership}{property manager of that site}
  \item{Anaplasma}{presence/absence of Anaplasma in a tick nymph from that site}
  \item{Babesia}{presence/absence of Babesia in a tick nymph from that site}
  \item{Lyme disease}{presence/absence of Lyme disease in a tick nymph that site}
  \item{Not infected}{Uninfected tick nymph found at that site}

}
}
\usage{
pathogens.table
}
\description{
Dataframe containing nymph tick abundances and presence/absence of tick-borne diseases.
}
\keyword{datasets}
