% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.R
\name{pie}
\alias{pie}
\title{Create a pie plot.}
\usage{
pie(df, prop, cat)
}
\arguments{
\item{df}{A DESCRIPTION.}

\item{prop}{A column in the 'df' dataframe. Must be of class 'numeric' and a proportion from 0-1.}

\item{cat}{A column in the 'df' dataframe. Must be of class factor or character.}
}
\value{
A ggplot2 rendering of the proportions across categories.
}
\description{
Create a pie plot.
}
\examples{
prop <- c(.2, .4, .4)
cat <- c("a", "b", "c")
mydata <- as.data.frame(cbind(prop, cat))
pie(mydata, mydata$prop, mydata$cat)
}
