% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plastic.sepa}
\alias{plastic.sepa}
\title{Plastic mulch & arthropod richness data}
\format{
R data.frame
\describe{
  \item{category}{factor of pitfall traps placed under plastic mulch ("Plastic) or not under plastic mulch ("No plastic")}
  \item{meanAbun}{mean abundance of arthropod predators in each pitfall trap}
  \item{meanRich}{mean family richness of arthropod predators in each pitfall trap}

}
}
\usage{
plastic.sepa
}
\description{
Dataframe containing arthropod abundance and richness data from traps based on presence/absence of plastic mulch.
}
\keyword{datasets}
