% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Create a scatter plot.}
\usage{
scatter(df, x, y, cat, xlab, ylab, legendLab)
}
\arguments{
\item{df}{DESCRIPTION.}

\item{x}{DESCRIPTION}

\item{y}{DESCRIPTION}

\item{cat}{DESCRIPTION.}

\item{xlab}{DESCRIPTION.}

\item{ylab}{DESCRIPTION.}

\item{legendLab}{DESCRIPTION.}
}
\value{
A ggplot2 rendering of DESCRIPTION.
}
\description{
Create a scatter plot.
}
\examples{
set.seed(123)
n <- 100
b0 <- 33
b1 <- 2
sigma <- 1
x <- round((rnorm(n, mean = 2, sd = .5)), 1)
y <- round((b0 + b1*x + rnorm(n, sigma)), 1)
cat <- c(rep("Group_A", 50), rep("Group_B", 50))
xlab <- "X-axis label"
ylab <- "Y-axis label"
legendLab <- "Legend Title"
mydata <- as.data.frame(cbind(as.numeric(x), as.numeric(y), cat))

scatter(mydata, x, y, cat, xlab, ylab, legendLab)
}
