% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tick.sites}
\alias{tick.sites}
\title{Tick and camera data availability by site}
\format{
R data.frame
\describe{
  \item{Site}{site name}
  \item{site_code}{site code (two-three letter unique identifier when not anonymized)}
  \item{Ownership}{property manager of that site}
  \item{Ticks}{presenece/absence of tick data that site}
  \item{Cameras}{presence/absence of camera trap pictures}
  \item{County}{county name}
}
}
\usage{
tick.sites
}
\description{
Dataframe containing arthropod abundance and richness data from traps based on presence/absence of plastic mulch.
}
\keyword{datasets}
