% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ticks.caught.county}
\alias{ticks.caught.county}
\title{Tick abundances by county data}
\format{
R data.frame
\describe{
  \item{County}{site county}
  \item{Species}{tick species' common name}
  \item{ticks county}{mean tick abundance}
  \item{sd_ticks_county}{standard deviation of tick abundance}

}
}
\usage{
ticks.caught.county
}
\description{
Dataframe containing ticks aggregated across all life stages by county.
}
\keyword{datasets}
