% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ticks.caught}
\alias{ticks.caught}
\title{Tick abundances by site data}
\format{
R data.frame
\describe{
  \item{Site}{site name}
  \item{Species}{tick species common name}
  \item{County}{site county}
  \item{Ownership}{property manager of that site}
  \item{ticks_site}{number of ticks found at that site}

}
}
\usage{
ticks.caught
}
\description{
Dataframe containing tick abundances aggregated across life stages by site.
}
\keyword{datasets}
