% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ticks.caught.table}
\alias{ticks.caught.table}
\title{Tick presence/absence data}
\format{
R data.frame
\describe{
  \item{Site}{site name}
  \item{County}{site county}
  \item{Ownership}{property manager of that site}
  \item{American Dog Tick}{presence/absence of American Dog Tick at that site}
  \item{Black Legged Tick}{presence/absence of Black Legged Tick at that site}
  \item{Brown Dog Tick}{presence/absence of Brown Dog Tick at that site}
  \item{Lone Star Tick}{presence/absence of Lone Star Tick at that site}
  \item{Longhorned Tick}{presence/absence of Longhorned Tick at that site}

}
}
\usage{
ticks.caught.table
}
\description{
Dataframe containing the presence/absence data of tick species for each site.
}
\keyword{datasets}
