% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{town}
\alias{town}
\title{Template data}
\format{
R data.frame
\describe{
\item{town}{name of town}
\item{park}{name of park within town}
\item{day}{day of sampling}
\item{num_trees}{number of trees sampled in the park}
\item{num_birds}{number of birds counted in the park on that day}
}
}
\usage{
town
}
\description{
Dataframe containing sample data from a fictitious survey of birds and trees from three towns
In this dataset, three parks, all named affectionately after insect groups, in each town were surveyed for ten days.
}
\keyword{datasets}
