% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{bar}
\alias{bar}
\title{Create a bar plot.}
\usage{
bar(df, x, y, ylab, fillLab)
}
\arguments{
\item{df}{A dataframe.}

\item{x}{A column in the 'df' dataframe. Must be of class 'numeric' and a proportion from 0-1.}

\item{y}{A column in the 'df' dataframe. Must be of class factor or character.}
}
\value{
A ggplot2 rendering of the proportions across categories.
}
\description{
Create a bar plot.
}
\examples{
x <- c("a", "b", "c")
y <- c(2, 4, 4)
mydata <- as.data.frame(cbind(x, y))
ylab <- c("y-axis label")
fillLab <- x
bar(df = mydata,
    x = mydata$x,
    y = mydata$y,
    ylab = ylab,
    fillLab = fillLab)
}
