% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curacao.exo.herps}
\alias{curacao.exo.herps}
\title{Curaçao exotic herps data}
\format{
R data.frame
\describe{
  \item{Site}{name of the focal property (anonymized by letter here)}
  \item{sitetype}{land-use category for the focal property}
  \item{Mourning Gecko}{abundance of Mourning Gecko at focal property.}
  \item{Tropical House Gecko}{abundance of Tropical House Gecko at focal property.}
  \item{Johnstone's Whistling Frog}{abundance of Johnstone's Whistling Frog at focal property}
  \item{Cuban Tree Frog}{abundance of Cuban Tree Frog at focal property}
  \item{Colombian Four-eyed Frog}{abundance of Colombian Four-eyed Frog at focal property}
  \item{Asian House Gecko}{abundance of Asian House Gecko at focal property}
  \item{ReportName}{name of report for references (not required)}

}
}
\usage{
curacao.exo.herps
}
\description{
Dataframe containing exotic reptile and amphibian data from Curaçao homeowner sampling.
}
\keyword{datasets}
