% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curacao.exo.herps.sum}
\alias{curacao.exo.herps.sum}
\title{curacao.exo.herps.sum}
\format{
\describe{
  \item{species}{Common name of exotic herp species observed on properties}
  \item{Site Type}{Land-use category for the focal property}
  \item{Percentage of Sites}{The number of sites where each herp species was observed in each category}

}
}
\usage{
curacao.exo.herps.sum
}
\description{
Some description.
}
\keyword{datasets}
