% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curacao.nat.birds.list}
\alias{curacao.nat.birds.list}
\title{List of Curaçao native birds}
\format{
R data.frame
\describe{
  \item{Site}{name of the focal property (anonymized by letter here)}
  \item{sitetype}{land-use category for the focal property}
  \item{Scientific name}{number of sites where each herp species was observed in each category}
  \item{Present}{number of sites where each herp species was observed in each category}
  \item{Common Name}{common name of native bird species in Curaçao}

}
}
\usage{
curacao.nat.birds.list
}
\description{
Dataframe with binary presence/absence variable for each bird species found at a Curaçao homeowner site.
}
\keyword{datasets}
