% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curacao.nat.birds}
\alias{curacao.nat.birds}
\title{curacao.nat.bird}
\format{
\describe{
  \item{Site}{Name of the focal property (anonymized by letter here)}
  \item{sitetype}{Land-use category for the focal property}
  \item{SR}{Species richness of native birds found at the focal property}
  \item{ReportName}{Name of report for references (not necessary)}

}
}
\usage{
curacao.nat.birds
}
\description{
Some description.
}
\keyword{datasets}
