% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curacao.nat.herps.sum}
\alias{curacao.nat.herps.sum}
\title{Curaçao native herp summary data}
\format{
R data.frame
\describe{
  \item{sitetype}{land-use category for the focal property}
  \item{species}{common name of native herp species on properties}
  \item{mean}{mean abundance of each native herp species on each property within a land-use category}

}
}
\usage{
curacao.nat.herps.sum
}
\description{
Dataframe summarizing the sites where native reptiles and amphibians were observed at Curaçao homeowner sites.
}
\keyword{datasets}
