% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.R
\name{pie}
\alias{pie}
\title{Create a pie plot.}
\usage{
pie(df, prop, cat)
}
\arguments{
\item{df}{A dataframe.}

\item{prop}{A column in the 'df' dataframe. Must be a proportion from 0-1.}

\item{cat}{A column in the 'df' dataframe. Must be be a factor.}
}
\value{
A ggplot rendering of the proportion of values in each category.
}
\description{
Create a pie plot.
}
\examples{
pie(mydat, mydat$prop, mydat$cat)
}
