% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Create a scatter plot.}
\usage{
scatter(df, x, y, cat, xlab, ylab, catName)
}
\arguments{
\item{df}{A dataframe.}

\item{cat}{A column in the 'df' dataframe. Must be of class factor or character.}

\item{prop}{A column in the 'df' dataframe. Must be of class 'numeric' and a proportion from 0-1.}
}
\value{
A ggplot2 rendering of the proportions across categories.
}
\description{
Create a scatter plot.
}
\examples{
set.seed(123)
n <- 100
b0 <- 33
b1 <- 2
sigma <- 1
x <- round((rnorm(n, mean = 2, sd = .5)), 1)
y <- round((b0 + b1*x + rnorm(n, sigma)), 1)
cat <- c(rep("a", 50), rep("b", 50))
xlab <- "X-axis label"
ylab <- "Y-axis label"
catName <- "Categories"
mydata <- as.data.frame(cbind(x, y, cat))
scatter(mydata, mydata$x, mydata$y, cat, xlab, ylab, catName)
}
